remove old pld
